+++ package.json
@@ -4,10 +4,10 @@
   "description": "A simple blockchain created using Javascript",
   "main": "index.js",
   "scripts": {
-    "tsc:": "tsc",
+    "tsc": "tsc",
     "tsc:watch": "tsc --w",
-    "start": "tsc && node dist/server.js",
-    "watch": "nodemon dist/server.js",
+    "start": "tsc && node dist/miner.js",
+    "watch": "nodemon dist/miner.js",
     "lint": "tslint --project tslint.json src/**/*.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -29,8 +29,7 @@
     "minimist": "^1.2.0",
     "tslint": "^5.7.0",
     "tslint-config-airbnb": "^5.3.0",
-    "typescript": "^2.5.3",
-    "uuid4": "^1.0.0"
+    "typescript": "^2.5.3"
   },
   "devDependencies": {
     "body-parser": "^1.18.2",
