+++ src/Block.ts
@@ -2,11 +2,11 @@
 import Transaction from './Transaction';
 
 interface Block {
-  readonly timestamp: number;
   readonly transactions: Array<Transaction>;
   readonly difficulty: number;
   readonly previousHash?: string;
-
+  
+  timestamp: number;
   nonce: number;
 }
